FIX: Ensure that custom header links migration do not fail validation (#55)
This commit is a follow up to 73747938bde3c2d3f6f68350c48a320364be1b04
where the migration will fail because the objects created by the
migration will end up failing to save because the objects will fail the
validation given the new schema. This commit updates the migration to
ensure that we do not end up with invalid objects.